pyproject.toml
[project]
name = "auraflow-ft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.5,<2.6",
    "torchvision>=0.20",
    "safetensors>=0.4.5",
    "hf-transfer>=0.1.8",
    "wandb>=0.19.1",
    "tqdm>=4.67.1",
    "pyyaml>=6.0.2",
    "lightning>=2.5.0.post0",
    "pydantic>=2.10.4",
    "transformers>=4.47.1",
    "datasets>=3.2.0",
    "webdataset>=0.2.100",
    "torchmetrics>=1.6.0",
    "bitsandbytes>=0.45.0",
    "schedulefree>=1.4",
    "diffusers>=0.32.0",
    "sentencepiece>=0.2.0",
    "accelerate>=1.2.1",
    "flash-attn>=2.7.2.post1",
    "torchao>=0.7.0",
    "optimum-quanto>=0.2.6",
    "triton>=3.1.0",
    "imagesize>=1.4.1",
    "cheesechaser>=0.2.2",
]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.1.1",
    "ruff>=0.8.4",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "pytest>=8.3.4",
    "fire>=0.7.0",
    "huggingface-dl",
    "click>=8.1.7",
    "litserve>=0.2.5",
    "gradio>=5.9.1",
]
no-build-isolation-package = ["flash-attn"]


[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
torch = [{ index = "pytorch-cu124", marker = "platform_system != 'Darwin'" }]
torchvision = [
    { index = "pytorch-cu124", marker = "platform_system != 'Darwin'" },
]
huggingface-dl = { git = "https://github.com/p1atdev/huggingface_dl" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
